Add more stubs to unit test better
