+++ src/test/java/seedu/address/logic/commands/TagCommandTest.java
@@ -1,27 +1,16 @@
 package seedu.address.logic.commands;
 
-import static java.util.Objects.requireNonNull;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static seedu.address.testutil.Assert.assertThrows;
 
-import java.nio.file.Path;
-import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.List;
-import java.util.function.Predicate;
 
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 
-import javafx.collections.ObservableList;
-import seedu.address.commons.core.GuiSettings;
 import seedu.address.logic.commands.exceptions.CommandException;
-import seedu.address.model.AddressBook;
-import seedu.address.model.Model;
-import seedu.address.model.ReadOnlyAddressBook;
-import seedu.address.model.ReadOnlyUserPrefs;
 import seedu.address.model.tag.Tag;
 import seedu.address.testutil.TagBuilder;
 
@@ -34,6 +23,7 @@ public class TagCommandTest {
 
     @Test
     @Disabled
+    @Deprecated
     // This test is using the original idea of tagging person.
     public void execute_tagAcceptedByModel_tagSuccessful() throws Exception {
         ModelStubAcceptingTagAdded modelStub = new ModelStubAcceptingTagAdded();
@@ -83,127 +73,5 @@ public class TagCommandTest {
         assertFalse(tagT1Command.equals(tagT2Command));
     }
 
-    /**
-     * A default model stub that have all of the methods failing.
-     */
-    private class ModelStub implements Model {
-        @Override
-        public void setUserPrefs(ReadOnlyUserPrefs userPrefs) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public ReadOnlyUserPrefs getUserPrefs() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public GuiSettings getGuiSettings() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setGuiSettings(GuiSettings guiSettings) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public Path getAddressBookFilePath() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setAddressBookFilePath(Path addressBookFilePath) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void addTag(Tag tag) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setAddressBook(ReadOnlyAddressBook newData) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public ReadOnlyAddressBook getAddressBook() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public boolean hasTag(Tag tag) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void deleteTag(Tag target) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setTag(Tag target, Tag editedTag) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public ObservableList<Tag> getFilteredTagList() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void updateFilteredTagList(Predicate<Tag> predicate) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public List<Tag> findFilteredTagList(Predicate<Tag> predicate) {
-            throw new AssertionError("This method should not be called.");
-        }
-    }
-
-    /**
-     * A Model stub that contains a single tag.
-     */
-    private class ModelStubWithTag extends ModelStub {
-        private final Tag tag;
-
-        ModelStubWithTag(Tag tag) {
-            requireNonNull(tag);
-            this.tag = tag;
-        }
-
-        @Override
-        public boolean hasTag(Tag tag) {
-            requireNonNull(tag);
-            return this.tag.isSameTag(tag);
-        }
-    }
-
-    /**
-     * A Model stub that always accept the tag being added.
-     */
-    @Deprecated
-    private class ModelStubAcceptingTagAdded extends ModelStub {
-        final ArrayList<Tag> tagsAdded = new ArrayList<>();
-
-        @Override
-        public boolean hasTag(Tag tag) {
-            requireNonNull(tag);
-            return tagsAdded.stream().anyMatch(tag::isSameTag);
-        }
-
-        @Override
-        public void addTag(Tag tag) {
-            requireNonNull(tag);
-            tagsAdded.add(tag);
-        }
-
-        @Override
-        public ReadOnlyAddressBook getAddressBook() {
-            return new AddressBook();
-        }
-    }
 
 }
